chore: Cleanup fixtures and functions. (#85)
Remove unused fixtures and functions
Move fixtures to be as close to tests as possible. 
Only shared fixtures and functions in conftest files
