+++ emulation_system/emulation_system/consts.py
@@ -16,6 +16,7 @@ ROOT_DIR = os.path.normpath(
     os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "..")
 )
 
+ROBOT_SERVER_DEFAULT_PORT = 31950
 DEFAULT_CONFIGURATION_FILE_PATH = f"{ROOT_DIR}/configuration.json"
 CONFIGURATION_FILE_LOCATION_VAR_NAME = "CONFIGURATION_FILE_LOCATION"
 DOCKERFILE_DIR_LOCATION = f"{ROOT_DIR}/docker/"
