+++ emulation_system/tests/compose_file_creator/conversion_logic/conftest.py
@@ -1,5 +1,9 @@
 """Conftest for conversion logic."""
-from typing import Any, Dict, List, cast
+from typing import (
+    Any,
+    Dict,
+    cast,
+)
 
 import py
 import pytest
@@ -18,6 +22,9 @@ from emulation_system.compose_file_creator.settings.images import (
     TemperatureModuleImages,
     ThermocyclerModuleImages,
 )
+from emulation_system.opentrons_emulation_configuration import (
+    OpentronsEmulationConfiguration,
+)
 from tests.compose_file_creator.conftest import (
     HEATER_SHAKER_MODULE_ID,
     MAGNETIC_MODULE_ID,
@@ -25,7 +32,6 @@ from tests.compose_file_creator.conftest import (
     TEMPERATURE_MODULE_ID,
     THERMOCYCLER_MODULE_ID,
 )
-from tests.conftest import get_test_conf
 
 CONTAINER_NAME_TO_IMAGE = {
     OT2_ID: RobotServerImages().local_firmware_image_name,
@@ -48,95 +54,25 @@ EMULATION_CONFIGURATION_FILE_NAME = "test-config.json"
 
 
 @pytest.fixture
-def version_only() -> Dict[str, Any]:
-    """Input file with only a compose-file-version specified."""
-    return {"compose-file-version": "4.0"}
-
-
-@pytest.fixture
-def extra_mounts_dir(tmpdir: py.path.local) -> str:
-    """Creates temp path for extra mounts."""
-    return str(tmpdir.mkdir("log_files"))
-
-
-@pytest.fixture
-def extra_mounts_and_opentrons(extra_mounts_dir: str, opentrons_dir: str) -> List[str]:
-    """List with mount dir and opentrons-modules dir together.
-
-    I have to build the list in a separate fixture because pytest-lazy-fixture does
-    not evaluate when the function is in a list.
-    So [lazy_fixture("opentrons"), lazy_fixture("extra_mounts_dir")]
-    does not evaluate to ["path/to/opentrons", "path/to/extra_mounts_dir"]
-    it instead evaluates to
-    ["<LazyFixture "opentrons_dir">, <LazyFixture "extra_mounts_dir">] which
-    causes the test to fail.
-    See issue here: https://github.com/TvoroG/pytest-lazy-fixture/issues/24
-    """
-    return [extra_mounts_dir, opentrons_dir]
-
-
-@pytest.fixture
-def modules_dir_in_list(opentrons_modules_dir: str) -> List[str]:
-    """Opentrons-modules repo in list.
-
-    See extra_mounts_and_opentrons_modules docstring for more info.
-    """
-    return [opentrons_modules_dir]
-
-
-@pytest.fixture
-def opentrons_dir_in_list(opentrons_dir: str) -> List[str]:
-    """Opentrons-modules repo in list.
-
-    See extra_mounts_and_opentrons_modules docstring for more info.
-    """
-    return [opentrons_dir]
-
-
-@pytest.fixture
-def robot_with_mount(ot2_only: Dict[str, Any], extra_mounts_dir: str) -> Dict[str, Any]:
-    """Robot dict with a mount added."""
+def robot_with_mount_and_modules_services(
+    tmpdir: py.path.local,
+    ot2_only: Dict[str, Any],
+    modules_only: Dict[str, Any],
+    testing_global_em_config: OpentronsEmulationConfiguration,
+) -> Dict[str, Service]:
+    """Get services from ot2_and_modules."""
     ot2_only["robot"]["extra-mounts"] = [
         {
             "name": "LOG_FILES",
             "type": MountTypes.DIRECTORY,
             "mount-path": EXTRA_MOUNT_PATH,
-            "source-path": extra_mounts_dir,
+            "source-path": str(tmpdir.mkdir("log_files")),
         }
     ]
-    return ot2_only
-
-
-@pytest.fixture
-def robot_with_mount_and_modules(
-    robot_with_mount: Dict[str, Any],
-    modules_only: Dict[str, Any],
-) -> Dict[str, Any]:
-    """Create config with robots and modules."""
-    robot_with_mount.update(modules_only)
-
-    return robot_with_mount
-
-
-@pytest.fixture
-def robot_with_mount_and_modules_services(
-    robot_with_mount_and_modules: Dict[str, Any],
-) -> Dict[str, Service]:
-    """Get services from ot2_and_modules."""
+    ot2_only.update(modules_only)
     return cast(
         Dict[str, Service],
-        convert_from_obj(robot_with_mount_and_modules, get_test_conf()).services,
-    )
-
-
-@pytest.fixture
-def with_system_unique_id_services(
-    with_system_unique_id: Dict[str, Any]
-) -> Dict[str, Service]:
-    """Get services from with_system_unique_id."""
-    return cast(
-        Dict[str, Service],
-        convert_from_obj(with_system_unique_id, get_test_conf()).services,
+        convert_from_obj(ot2_only, testing_global_em_config).services,
     )
